address review 2 from easwar
