+++ xds/internal/balancer/cdsbalancer/cdsbalancer_security_test.go
@@ -140,7 +140,7 @@ func setupForSecurityTests(t *testing.T, bootstrapContents []byte, clientCreds, 
 
 	config, err := bootstrap.NewConfigForTesting(bootstrapContents)
 	if err != nil {
-		t.Fatalf("Failed to create an bootstrap config from contents: %v, %v", bootstrapContents, err)
+		t.Fatalf("Failed to parse bootstrap contents: %s, %v", string(bootstrapContents), err)
 	}
 	pool := xdsclient.NewPool(config)
 	xdsClient, xdsClose, err := pool.NewClientForTesting(xdsclient.OptionsForTesting{
