+++ xds/internal/xdsclient/tests/ads_stream_flow_control_test.go
@@ -152,7 +152,7 @@ func createXDSClient(t *testing.T, bootstrapContents []byte) xdsclient.XDSClient
 
 	config, err := bootstrap.NewConfigForTesting(bootstrapContents)
 	if err != nil {
-		t.Fatalf("Failed to create an bootstrap config from contents: %v, %v", bootstrapContents, err)
+		t.Fatalf("Failed to parse bootstrap contents: %s, %v", string(bootstrapContents), err)
 	}
 	pool := xdsclient.NewPool(config)
 	client, close, err := pool.NewClientForTesting(xdsclient.OptionsForTesting{
